Silence Visual C++ 8.0 warnings about data loss as a result of type conversion.
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@17502 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ juggler/modules/vapor/vpr/IO/Socket/SocketBasicOpt.h
@@ -123,7 +123,7 @@ public:
    /**
     *
     */
-   vpr::ReturnStatus getRecvBufferSize(Int32& size)
+   vpr::ReturnStatus getRecvBufferSize(size_t& size)
    {
       vpr::SocketOptions::Data opt;
       vpr::ReturnStatus retval;
@@ -157,7 +157,7 @@ public:
    /**
     *
     */
-   vpr::ReturnStatus getSendBufferSize(int& size)
+   vpr::ReturnStatus getSendBufferSize(size_t& size)
    {
       vpr::SocketOptions::Data opt;
       vpr::ReturnStatus retval;
